Improved coverage

+++ aw_datastore/storages/mongodb.py
@@ -46,8 +46,13 @@ class MongoDBStorage(AbstractStorage):
         self.db[bucket_id]["metadata"].insert_one(metadata)
 
     def delete_bucket(self, bucket_id: str) -> None:
-        self.db[bucket_id]["events"].drop()
-        self.db[bucket_id]["metadata"].drop()
+        print(self.db.collection_names())
+        if bucket_id + ".metadata" in self.db.collection_names():
+            self.db[bucket_id]["events"].drop()
+            self.db[bucket_id]["metadata"].drop()
+        else:
+            # TODO: Create custom exception
+            raise Exception('Bucket did not exist, could not delete')
 
     def buckets(self) -> Dict[str, dict]:
         bucketnames = set()
@@ -63,7 +68,9 @@ class MongoDBStorage(AbstractStorage):
         metadata = self.db[bucket_id]["metadata"].find_one({"_id": "metadata"})
         if metadata:
             del metadata["_id"]
-        return metadata
+            return metadata
+        else:
+            raise Exception('Bucket did not exist, could not get metadata')
 
     def get_events(self, bucket_id: str, limit: int,
                    starttime: Optional[datetime] = None, endtime: Optional[datetime] = None):
@@ -91,7 +98,7 @@ class MongoDBStorage(AbstractStorage):
         return events
 
     def get_eventcount(self, bucket_id: str,
-                   starttime: Optional[datetime] = None, endtime: Optional[datetime] = None):
+                       starttime: datetime = None, endtime: datetime = None) -> int:
         query_filter = {}  # type: Dict[str, dict]
         if starttime or endtime:
             query_filter["timestamp"] = {}
@@ -102,7 +109,7 @@ class MongoDBStorage(AbstractStorage):
         return self.db[bucket_id]["events"].find(query_filter).count()
 
     def _transform_event(self, event: dict) -> dict:
-        if "duration" in event:
+        if "duration" in event:  # pragma: no cover
             event["duration"] = event["duration"].total_seconds()
         return event
